feat: full refactor from an SPA like site structure
