+++ backend/package.json
@@ -8,7 +8,7 @@
     "dev": "nodemon server.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "lint": "eslint .",
-    "postinstall": "npm rebuild sharp --platform=linux --arch=x64",
+    "rebuild-sharp-lambda": "npm rebuild sharp --platform=linux --arch=x64",
     "cleanup-aws": "node scripts/cleanup-aws-credentials.js"
   },
   "dependencies": {
@@ -20,7 +20,7 @@
     "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
-    "express": "^4.22.1",
+    "express": "^4.22.2",
     "express-jwt": "^8.5.1",
     "express-rate-limit": "^7.5.1",
     "express-validator": "^7.3.1",
